# packages/shared-lib/src/app/openapi/easy-genomics-api.yaml
openapi: 3.1.0
info:
  title: Easy Genomics API
  version: 1.0.0
  description: |-
    Internal REST API for the Easy Genomics platform.
    All routes require a Cognito JWT bearer token unless marked with security: [].
  license:
    name: Proprietary
servers:
  - url: https://api.example.com
    description: Easy Genomics API (placeholder — replaced by CDK integration)
components:
  securitySchemes:
    cognitoJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cognito User Pool JWT token
  schemas:
    ErrorResponse:
      type: object
      properties:
        Error:
          type: string
        ErrorCode:
          type: string
          description: >-
            EG-1xx: generic, EG-2xx: organization, EG-3xx: laboratory, EG-4xx: user, EG-5xx: AWS HealthOmics, EG-6xx:
            Seqera Cloud
      required:
        - Error
        - ErrorCode
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
security:
  - cognitoJwt: []
paths:
...
  /easy-genomics/data-collections/list-sequence-sets-by-tag:
    get:
      operationId: listSequenceSetsByTag
      summary: List Sequence Sets By Tag
      parameters:
        - name: laboratoryId
          in: query
          required: true
          schema:
            type: string
          description: Laboratory to query
        - name: tagId
          in: query
          required: true
          schema:
            type: string
          description: Tag to list sequence sets for
        - name: limit
          in: query
          required: false
          schema:
            type: string
          description: Max number of results (1-500)
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Pagination cursor
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...